DXTerrain checkpoint 010426

+++ src/DXTerrain/Tools.h
@@ -13,7 +13,7 @@ namespace NSTool
 
         int size = std::snprintf(nullptr, 0, format.c_str(), to_str(args)...);
         if (size < 0) return {};
-        std::vector<char> buf(size + 1);
+        std::vector<char> buf(static_cast<size_t>(size + 1));
         std::snprintf(buf.data(), buf.size(), format.c_str(), to_str(args)...);
         return std::string(buf.data(), size);
     }
@@ -28,7 +28,7 @@ namespace NSTool
 
         int size = std::swprintf(nullptr, 0, format.c_str(), to_wstr(args)...);
         if (size < 0) return {};
-        std::vector<wchar_t> buf(size + 1);
+        std::vector<wchar_t> buf(static_cast<size_t>(size + 1));
         std::swprintf(buf.data(), buf.size(), format.c_str(), to_wstr(args)...);
         return std::wstring(buf.data(), size);
     }
